+++ src/main/java/com/github/legoatoom/goldenberries/blocks/GoldenBerryBushBlock.java
@@ -22,11 +22,9 @@ import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
 import net.minecraft.block.SweetBerryBushBlock;
-import net.minecraft.client.sound.SoundEngine;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
-import net.minecraft.server.world.ServerWorld;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvents;
 import net.minecraft.tag.BlockTags;
@@ -37,15 +35,12 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.BlockView;
 import net.minecraft.world.World;
 
-import java.util.Random;
-
 /**
  * The same as a normal SweetBerryBushBlock only that I had to overwrite some methods that dropped sweet_berries.
  *
  * @author legoatoom
  */
 public class GoldenBerryBushBlock extends SweetBerryBushBlock {
-
     public GoldenBerryBushBlock(Settings settings) {
         super(settings);
         this.setDefaultState(this.stateManager.getDefaultState().with(AGE, 0));
@@ -57,20 +52,21 @@ public class GoldenBerryBushBlock extends SweetBerryBushBlock {
     }
 
     @Override
-    protected boolean canPlantOnTop(BlockState floor, BlockView world, BlockPos pos) {
-        return floor.isIn(BlockTags.SAND) || floor.isOf(Blocks.NETHERRACK)
-                || floor.isIn(BlockTags.NYLIUM) || floor.isOf(Blocks.SOUL_SOIL) || super.canPlantOnTop(floor, world, pos);
-    }
-
-    @Override
     public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
-        int i = state.get(AGE);
-        boolean isFullGrown = i == 3;
+        int     ageInt      = state.get(AGE);
+        boolean isFullGrown = ageInt == 3;
         if (!isFullGrown && player.getStackInHand(hand).isOf(Items.BONE_MEAL)) {
             return ActionResult.PASS;
         }
-        if (i > 1) {
-            int j = 1 + world.random.nextInt(2);
+        if (ageInt > 1) {
+            int j = world.random.nextInt(2);
+            if (world.getRegistryKey() == World.NETHER) {
+                int nuggetDrop = world.random.nextInt(4);
+                if (nuggetDrop == 0) {
+                    nuggetDrop = world.random.nextInt(3);
+                    SweetBerryBushBlock.dropStack(world, pos, new ItemStack(Items.GOLD_NUGGET, nuggetDrop + (isFullGrown ? 1 : 0)));
+                }
+            }
             SweetBerryBushBlock.dropStack(world, pos, new ItemStack(ModItems.GOLDEN_BERRIES, j + (isFullGrown ? 1 : 0)));
             //Possibly use a different sound?
             world.playSound(null, pos, SoundEvents.BLOCK_SWEET_BERRY_BUSH_PICK_BERRIES, SoundCategory.BLOCKS, 1.0f, 0.8f + world.random.nextFloat() * 0.4f);
@@ -78,5 +74,10 @@ public class GoldenBerryBushBlock extends SweetBerryBushBlock {
             return ActionResult.success(world.isClient);
         }
         return super.onUse(state, world, pos, player, hand, hit);
+    }
+
+    @Override
+    protected boolean canPlantOnTop(BlockState floor, BlockView world, BlockPos pos) {
+        return floor.isIn(BlockTags.SAND) || floor.isOf(Blocks.NETHERRACK) || floor.isIn(BlockTags.NYLIUM) || floor.isOf(Blocks.SOUL_SOIL) || super.canPlantOnTop(floor, world, pos);
     }
 }