+++ src/main/java/com/github/legoatoom/goldenberries/blocks/ModBlocks.java
@@ -19,7 +19,8 @@ package com.github.legoatoom.goldenberries.blocks;
 
 import com.github.legoatoom.goldenberries.GoldenBerries;
 import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
-import net.minecraft.block.*;
+import net.minecraft.block.Block;
+import net.minecraft.block.Material;
 import net.minecraft.sound.BlockSoundGroup;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.Registry;
@@ -32,13 +33,16 @@ import net.minecraft.util.registry.Registry;
 @SuppressWarnings("SameParameterValue")
 public class ModBlocks {
 
-    public static Block GOLDEN_BERRY_BUSH = new GoldenBerryBushBlock(FabricBlockSettings.of(Material.PLANT).ticksRandomly().noCollision().sounds(BlockSoundGroup.SWEET_BERRY_BUSH));
+    public static Block GOLDEN_BERRY_BUSH = new GoldenBerryBushBlock(FabricBlockSettings.of(Material.PLANT)
+                                                                                        .ticksRandomly()
+                                                                                        .noCollision()
+                                                                                        .sounds(BlockSoundGroup.SWEET_BERRY_BUSH));
 
     private static void register(String id, Block block) {
         Registry.register(Registry.BLOCK, new Identifier(GoldenBerries.MOD_ID, id), block);
     }
 
-    public static void init(){
+    public static void init() {
         register("golden_berry_bush", GOLDEN_BERRY_BUSH);
     }
 }