+++ src/main/java/com/github/legoatoom/goldenberries/mixin/client/gui/hud/InGameHudMixin.java
@@ -36,40 +36,45 @@ import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 import static net.minecraft.client.gui.hud.InGameHud.HeartType.*;
 
 /**
- * Mixin class for the {@link InGameHud} that creates the blue hearts when the user has the potion effect on.
- * This class switches what texture the renderer should use to render the hearts.
- * Because I can't directly locate where the code should begin and stop I had to use two injects to find a start and
- * endpoint to turn off the texture and replace the original texture.
+ * Mixin class for the {@link InGameHud} that creates the blue hearts when the user has the potion effect on. This class
+ * switches what texture the renderer should use to render the hearts. Because I can't directly locate where the code
+ * should begin and stop I had to use two injects to find a start and endpoint to turn off the texture and replace the
+ * original texture.
  *
  * @author legoatoom
  */
 @Mixin(InGameHud.class)
 public abstract class InGameHudMixin extends DrawableHelper {
 
-    private static final Identifier MOD_GUI_ICONS_TEXTURE = new Identifier(GoldenBerries.MOD_ID,"textures/gui/icons.png");
+    private static final Identifier MOD_GUI_ICONS_TEXTURE = new Identifier(GoldenBerries.MOD_ID, "textures/gui/icons.png");
 
 
     private int heartX;
     private int heartY;
     private int heartIndex;
 
-    @Shadow protected abstract void drawHeart(MatrixStack matrices, InGameHud.HeartType type, int x, int y, int v, boolean blinking, boolean halfHeart);
+    @Shadow
+    protected abstract void drawHeart(MatrixStack matrices, InGameHud.HeartType type, int x, int y, int v, boolean blinking, boolean halfHeart);
 
-    @Inject(method = "renderHealthBar", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/hud/InGameHud;drawHeart(Lnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/gui/hud/InGameHud$HeartType;IIIZZ)V", ordinal = 0), locals = LocalCapture.CAPTURE_FAILHARD)
+    @Inject(method = "renderHealthBar",
+            at = @At(value = "INVOKE",
+                     target = "Lnet/minecraft/client/gui/hud/InGameHud;drawHeart(Lnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/gui/hud/InGameHud$HeartType;IIIZZ)V",
+                     ordinal = 0),
+            locals = LocalCapture.CAPTURE_FAILHARD)
     private void storeLocals(MatrixStack matrices, PlayerEntity player, int x, int y, int lines, int regeneratingHeartIndex, float maxHealth, int lastHealth, int health, int absorption, boolean blinking, CallbackInfo ci, InGameHud.HeartType type, int i, int j, int k, int l, int m, int n, int o, int p, int q) {
-        this.heartX = p;
-        this.heartY = q;
+        this.heartX     = p;
+        this.heartY     = q;
         this.heartIndex = m;
     }
 
-    @Inject(method = "renderHealthBar", at =
-    @At(value = "INVOKE",
-            target = "Lnet/minecraft/client/gui/hud/InGameHud;drawHeart(Lnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/gui/hud/InGameHud$HeartType;IIIZZ)V",
-            ordinal = 3,
-            shift = At.Shift.BY,
-            by = 2))
+    @Inject(method = "renderHealthBar",
+            at = @At(value = "INVOKE",
+                     target = "Lnet/minecraft/client/gui/hud/InGameHud;drawHeart(Lnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/gui/hud/InGameHud$HeartType;IIIZZ)V",
+                     ordinal = 3,
+                     shift = At.Shift.BY,
+                     by = 2))
     private void renderOverlay(MatrixStack matrices, PlayerEntity player, int x, int y, int lines, int regeneratingHeartIndex, float maxHealth, int lastHealth, int health, int absorption, boolean blinking, CallbackInfo ci) {
-        if (player.hasStatusEffect(ModStatusEffects.POISON_RESISTANCE)){
+        if (player.hasStatusEffect(ModStatusEffects.POISON_RESISTANCE)) {
             RenderSystem.setShaderTexture(0, MOD_GUI_ICONS_TEXTURE);
             drawResistanceHearts(matrices, player, lastHealth, health, blinking);
             RenderSystem.setShaderTexture(0, GUI_ICONS_TEXTURE);
@@ -78,13 +83,14 @@ public abstract class InGameHudMixin extends DrawableHelper {
 
     /**
      * Does all the drawing of the hearts.
+     *
      * @see InGameHud#renderHealthBar(MatrixStack, PlayerEntity, int, int, int, int, float, int, int, int, boolean)
      */
     private void drawResistanceHearts(MatrixStack matrices, PlayerEntity player, int lastHealth, int health, boolean blinking) {
-        InGameHud.HeartType heartType = player.hasStatusEffect(StatusEffects.POISON) ? POISIONED : (player.hasStatusEffect(StatusEffects.WITHER) ? WITHERED : (player.isFreezing() ? FROZEN : NORMAL));
-        int i = 9 * (player.world.getLevelProperties().isHardcore() ? 5 : 0);
-        int heartIndex2 = heartIndex * 2;
-        boolean halfHeart;
+        InGameHud.HeartType heartType   = player.hasStatusEffect(StatusEffects.POISON) ? POISIONED : (player.hasStatusEffect(StatusEffects.WITHER) ? WITHERED : (player.isFrozen() ? FROZEN : NORMAL));
+        int                 i           = 9 * (player.world.getLevelProperties().isHardcore() ? 5 : 0);
+        int                 heartIndex2 = heartIndex * 2;
+        boolean             halfHeart;
         if (blinking && heartIndex2 < health) {
             halfHeart = heartIndex2 + 1 == health;
             this.drawHeart(matrices, heartType, heartX, heartY, i, true, halfHeart);