+++ src/main/java/com/github/legoatoom/goldenberries/mixin/entity/effect/StatusEffectInstanceMixin.java
@@ -1,0 +1,59 @@
+/*
+ * This file is part of GoldenBerries.
+ *
+ * GoldenBerries is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * GoldenBerries is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with GoldenBerries.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+/*
+ * This file is part of GoldenBerries.
+ *
+ * GoldenBerries is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * GoldenBerries is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with GoldenBerries.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.github.legoatoom.goldenberries.mixin.entity.effect;
+
+import com.github.legoatoom.goldenberries.entity.effect.StatusEffectTimeOut;
+import net.minecraft.entity.effect.StatusEffectInstance;
+import org.spongepowered.asm.mixin.Mixin;
+
+@Mixin(StatusEffectInstance.class)
+public class StatusEffectInstanceMixin implements StatusEffectTimeOut {
+    protected int timeOut = 0;
+
+    @Override
+    public int getTimeOut() {
+        return timeOut;
+    }
+
+    @Override
+    public void setTimeout(int val) {
+        this.timeOut = val;
+    }
+
+    @Override
+    public void reduceTimeOut() {
+        this.timeOut = Math.max(0, this.timeOut - 1);
+    }
+}