+++ changelog.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -6,18 +7,38 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Time is following the [Holocene Calendar](https://en.wikipedia.org/wiki/Holocene_calendar).
 
+## [1.2.3] - 12022-05-18
+
+### Changed
+
+- Golden Berry Bushes drop between 0 - 1 (+1 if fully grown) golden berries.
+- Potion of Alexiteric only stop damage in Peaceful/Easy, in Normal/Hard they increase the amount of time between
+  getting damaged.
+    - Alexiteric potions now have a strong variant that last 90 seconds, made with glowstone dust.
+    - Wither Roses cannot be stopped, they are mean and prick, a potion can't stop that.
+
+### Added
+
+- Golden Berry Bushes have an additional 1/4 chance to drop 0 - 3 (+1 if fully grown)
+  golden nuggets if planted in the Nether.
+
 ## [1.2.2] - 12021-12-11
+
 ### HotFix
+
 - Brewing Recipe not working
 
 ## [1.2.1] - 12021-12-11
+
 ### Fixed
+
 - Loading Bug causing crashes.
 - Piglins not loving _Golden Berries_
 
 ## [1.2.0] - 12021-12-09
 ### Added
-- _Golden Berry Bush_, very simmilar to _Sweet Berry Bush_
+
+- _Golden Berry Bush_, very similar to _Sweet Berry Bush_
 
 ### Changed
 - Duration of potion increased